requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.5.4
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
whitenoise==6.11.0
